# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Ges_Compte Laravel",
        "description": "API pour la gestion des comptes bancaires",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/comptes/{compte}": {
            "get": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Afficher un compte spécifique",
                "description": "Récupère les détails d'un compte bancaire spécifique",
                "operationId": "acb7ab8e6d93dab2d01f1ba8eb6e0e18",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du compte récupérés avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "nullable": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/CompteResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Mettre à jour un compte",
                "description": "Met à jour les informations d'un compte bancaire existant",
                "operationId": "5293c1a3f2cd81cb7294db8426c64ac7",
                "parameters": [
                    {
                        "name": "compte",
                        "in": "path",
                        "description": "ID du compte",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "titulaire": {
                                        "type": "string",
                                        "maxLength": 150,
                                        "example": "Jean Dupont"
                                    },
                                    "informationsUser": {
                                        "properties": {
                                            "nom": {
                                                "type": "string",
                                                "maxLength": 50,
                                                "example": "Dupont"
                                            },
                                            "prenom": {
                                                "type": "string",
                                                "maxLength": 70,
                                                "example": "Jean"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jean.dupont@email.com"
                                            },
                                            "telephone": {
                                                "type": "string",
                                                "example": "+221771234567"
                                            },
                                            "nci": {
                                                "type": "string",
                                                "example": "1234567890123456"
                                            },
                                            "adresse": {
                                                "type": "string",
                                                "example": "Dakar, Sénégal"
                                            },
                                            "password": {
                                                "type": "string",
                                                "minLength": 6,
                                                "example": "newpassword123"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Compte mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte mis à jour avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "format": "uuid"
                                                },
                                                "numeroCompte": {
                                                    "type": "string"
                                                },
                                                "titulaire": {
                                                    "type": "string"
                                                },
                                                "type": {
                                                    "type": "string"
                                                },
                                                "solde": {
                                                    "type": "number"
                                                },
                                                "devise": {
                                                    "type": "string"
                                                },
                                                "dateCreation": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "statut": {
                                                    "type": "string"
                                                },
                                                "metadata": {
                                                    "properties": {
                                                        "derniereModification": {
                                                            "type": "string",
                                                            "format": "date-time"
                                                        },
                                                        "version": {
                                                            "type": "integer"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Données de validation invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "CompteResource": {
                "title": "CompteResource",
                "description": "Ressource représentant un compte bancaire avec ses métadonnées",
                "properties": {
                    "id": {
                        "description": "ID unique du compte",
                        "type": "string",
                        "format": "uuid"
                    },
                    "numeroCompte": {
                        "description": "Numéro unique du compte",
                        "type": "string"
                    },
                    "titulaire": {
                        "description": "Nom du titulaire du compte",
                        "type": "string"
                    },
                    "type": {
                        "description": "Type de compte",
                        "type": "string",
                        "enum": [
                            "epargne",
                            "cheque"
                        ]
                    },
                    "solde": {
                        "description": "Solde du compte",
                        "type": "number",
                        "format": "decimal"
                    },
                    "devise": {
                        "description": "Devise du compte",
                        "type": "string"
                    },
                    "dateCreation": {
                        "description": "Date de création du compte",
                        "type": "string",
                        "format": "date-time"
                    },
                    "statut": {
                        "description": "Statut du compte",
                        "type": "string",
                        "enum": [
                            "actif",
                            "bloque",
                            "ferme"
                        ]
                    },
                    "motifBlocage": {
                        "description": "Motif de blocage si applicable",
                        "type": "string",
                        "nullable": true
                    },
                    "metadata": {
                        "description": "Métadonnées du compte",
                        "properties": {
                            "derniereModification": {
                                "description": "Date de dernière modification",
                                "type": "string",
                                "format": "date-time"
                            },
                            "version": {
                                "description": "Version du compte",
                                "type": "integer"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}